Add patches to remove tekit keyword from the scripts

See merge request ftp/tekit_backend!167

+++ scripts/package_trt_llm_backend.sh
@@ -26,18 +26,23 @@ rm -rf tests
 
 rm -rf tools/dataset
 
-rm -rf ci
-
-rm tools/gen_trtllm_dockerfile.py
-
 #
 # Closed-source Batch Manager changes
 #
 
-# patch the cmake to use lib with pre_cxx11.a suffix and remove the patch from release pacakge
+# patch the cmake to use lib with pre_cxx11.a suffix and remove the patch from release package
 patch inflight_batcher_llm/CMakeLists.txt  < inflight_batcher_llm/CMakeLists.txt.patch
 rm inflight_batcher_llm/CMakeLists.txt.patch
 [[ -e inflight_batcher_llm/CMakeLists.txt.orig ]] && rm inflight_batcher_llm/CMakeLists.txt.orig
+
+# patch the scripts to point to the TensorRT-LLM GitHub repo and remove the patch from release package
+patch ci/L0_backend_trtllm/generate_engines.sh  < ci/L0_backend_trtllm/generate_engines.sh.patch
+rm ci/L0_backend_trtllm/generate_engines.sh.patch
+[[ -e ici/L0_backend_trtllm/generate_engines.sh.orig ]] && rm ci/L0_backend_trtllm/generate_engines.sh.orig
+
+patch tools/gen_trtllm_dockerfile.py  < tools/gen_trtllm_dockerfile.py.patch
+rm tools/gen_trtllm_dockerfile.py.patch
+[[ -e tools/gen_trtllm_dockerfile.py.orig ]] && rm tools/gen_trtllm_dockerfile.py.orig
 
 # exit if the keyword is found
 grep "__LUNOWUD" -R > /dev/null && exit 1
